requirements.txt
crewai==0.74.2
requests==2.32.3
urllib3==2.2.3
openai==1.52.1
python-dotenv==1.0.1



